requirements.txt
grpcio==1.48.1
grpcio-tools==1.48.1
Pillow==9.4.0
protobuf==3.19.5
python-dotenv==1.0.0
six==1.16.0
stability-sdk==0.3.2
